Last phase of OOM refactoring. - Replaced BSGOutOfMemoryWatchdog with BugsnagSystemState. - Moved OOM detection into BugsnagClient, which uses BugsnagSystemState as its information source. - Stronger and more explicit OOM heuristics. - OOM detection now relies entirely upon last launch state (includng debugger status).
Under the new system, you can cause an out of memory shutdown while the app is not being debugged, and then launch it in a debugger to observe it sending an OOM event. Any sudden shutdowns while a debugger is active won't generate an OOM event (as before).

We should still add more heuristics later. All heuristics are contained and explained in a single easy-to-follow method "didLikelyOOM" in BugsnagClient.